Вносит правки после код-ревью
