+++ js/main.js
@@ -1,7 +1,7 @@
 import { makePhotos } from './create-array-photos.js';
 import { renderThumbnails } from './thumbnails.js';
 import { initPreview } from './render-photo.js';
-import { initUploadModal } from './upload-photo-form.js';
+import { initUploadModal } from './form/upload-photo-form.js';
 
 const photos = makePhotos();
